refactor: 유닛 테스트 패키지 구조를 프로덕션 코드에 맞게 변경
