+++ app/src/test/java/woowacourse/movie/ui/home/MovieHomePresenterTest.kt
@@ -1,16 +1,10 @@
-package woowacourse.movie.model.ui.home
+package woowacourse.movie.ui.home
 
-import androidx.room.Room
-import androidx.test.core.app.ApplicationProvider
 import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.movie.data.database.MovieDatabase
-import woowacourse.movie.data.database.movie.MovieContentDao
-import woowacourse.movie.model.ui.repository.FakeMovieContentRepository
-import woowacourse.movie.ui.home.MovieHomeContract
-import woowacourse.movie.ui.home.MovieHomePresenter
+import woowacourse.movie.ui.repository.FakeMovieContentRepository
 
 class MovieHomePresenterTest {
     private lateinit var presenter: MovieHomePresenter
