+++ app/src/test/java/woowacourse/movie/ui/repository/FakeTicketRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model.ui.repository
+package woowacourse.movie.ui.repository
 
 import woowacourse.movie.data.database.ticket.TicketDao
 import woowacourse.movie.data.database.ticket.TicketEntity
